telemetry: move wasmplugin e2e to api
