+++ tests/integration/telemetry/api/testdata/registry-secret.yaml
