+++ tests/integration/telemetry/api/testdata/wasm-filter.yaml
